mm/address_space: avoid double initialization of kernel mapping
ImmutAfterInitCell states that the init() method should not be called
on an already initialized instance, but KERNEL_MAPPING is initialized
via ImmutAfterInitCell::new() first, and then via the aforementioned
init() method. At the moment this requirement is not enforced in any
way,(ImmutAfterInitCell::reinit() just calls init()), but this will
not hold in the future.

Create the global structure with ImmutAfterInitCell::uninit(), since
there should be no users of the mapping before the call to
init_kernel_mapping_info(), which happens very early in the boot
process for both the stage 2 and the proper SVSM.

Signed-off-by: Carlos López <carlos.lopez@suse.com>