sev/vmsa: introduce VmsaPage type
Introduce a new type to abstract away VMSA memory management. The new
type is a wrapper around the PageBox type with some extra
functionality. This obsoletes the VmsaRef type.

While we are at it, clean up borrows in start_cpu(). Before, we were
holding on to a raw VmsaRef while the actual VMSA was being updated
inside PerCpu::prepare_svsm_vmsa(), which was not unsafe but
definitely an antipattern in Rust. These changes also obsolete the
PerCpu::get_svsm_vmsa() method.

Signed-off-by: Carlos López <carlos.lopez@suse.com>