sev/hv_doorbell: introduce the HVDoorbellPage type
Introduce a new type to abstract away the memory management of the HV
doorbell page. The new type is a wrapper around a PageBox with some
extra functionality.

While we are at it, since `HVDoorBell::process_pending_events()` takes
self, pass const pointers instead of mut pointers to the HV page.

Signed-off-by: Carlos López <carlos.lopez@suse.com>