build: do not hardcode host toolchain
Currently, the default target triple for the whole workspace is set
to `x86_64-unknown-none`, which means any workspace package that needs
to be built with the host target (e.g. host utilities) needs it
explicitly specified (e.g. `x86_64-unknown-linux-gnu`).

Instead, do not specify a default target triple, as the host tools
should be buildable from any system (e.g. non-x86), and explicitly
specify the bare-metal target triple for those packages that require
it (stage 1, stage 2 and the main SVSM kernel).

This also has the advantage of simplifying some Makefile targets, as
most workspace members are meant to run on the host anyway.

Signed-off-by: Carlos López <carlos.lopezr4096@gmail.com>