mm: introduce mapping access infrastructure
In the SVSM kernel there are multiple uses of types like
PerCPUPageMappingGuard, GuestPtr, SharedBox, etc., all of which aim to
grant access to memory mapped under certain conditions (memory being
local to the SVSM kernel, shared with the guest or shared with the host
hypervisor), and which correspond to different cautions one must take.

On top of that, these types handle ownership of the mappings in
different ways - for example, GuestPtr does not own the underlying
mapping, but has no notion of the validity of the mapping, which seems
problematic since we often construct GuestPtrs from raw virtual
addresses coming from a PerCPUPageMappingGuard.

All of these uses place a burden on the developer, since safety comments
for all these unsafe actions need to be very explicit about all the
assumptions present. To improve this, introduce new types to unify all
of these use cases under the same semantics, allowing for compile-time
dispatching of different access types, adding much needed safety, while
allowing flexibility when it comes to working with memory (e.g. dealing
with dynamically-shared structures).

This patch only introduces the basic infrastructure (types and traits)
for the mapping access framework. The following commits introduce
further specializations based on the type of memory being accessed, and
then progressively replace the intended users with the new API.

The API is made up of several components: at the lowest level, several
types are introduced to handle the low level details of reading and
writing from memory; these help dispatching access operations to the
right low-level function (e.g. GuestPtr wants to use do_movsb(), while
host-shared mappings want to use unsafe_copy_bytes(), etc.).

Next, we introduce the Mapping and MappingRef types. The former is a
typed version of PerCPUPageMappingGuard, which allows mapping physical
memory on demand, and unmapping it on drop. MappingRef on the other
hand does not own the virtual mapping, but has a notion of a lifetime,
meaning instances of MappingRef are typically valid only for the
lifetime of a Mapping.

Both mapping types have a common interface through several traits that
allow reading and writing the backing memory and treating it both as
sized and unsized (i.e. dynamic slices) types. Reading and writing are
always bound by the FromBytes and IntoBytes traits from the zerocopy
crate, completely disallowing types with invalid representations,
removing a lot of unsafe code. Moreover, both of these types perform all
the necesary bounds checking, since they have a notion of the size of
the mapping and the size of the backing type. Mapping and MappingRef are
also generic over the access types (Guest, Local, Shared, User), meaning
that the raw access to memory is compile-time dispatched.

The trait interface is flexible and also generic, meaning that one may
specify things like the following in function signatures:

    #[derive(FromBytes)
    struct MyType { ... }

    impl MyType {
        fn check(&self) -> Result<Self, SvsmError> { ... }
    }

    fn read_and_check<M>(mapping: M) -> Result<T, SvsmError>
    where
        M: ReadableMapping<Guest, MyType>,
    {
        m.read().and_then(|s| s.check())
    }

read_and_check() here asks for any kind of guest mapping (owned or
borrowed) to a MyStruct, and promises to only read (and not write) to
it. The low-level details (like accessing to memory, checking whether
the mapping is big enough and properly aligned to hold a MyStruct, etc.)
are handled by the `M` type.

As mentioned before, these types also allow handling the backing memory
in a flexible way. All mappings implement methods to reinterpret the
backing memory as a different type (e.g. a slice of bytes to zero it
out), again performing the right bound checks, allowing the public
interface to be safe.

Signed-off-by: Carlos López <clopez@suse.de>