mm/alloc: introduce SlabCommon::allocate_page_slot()
There are two users of SlabCommon::allocate_slot() which use the new
slot to store a SlabPage: the SlabPageSlab and Slab. Add a new method
for these two users to reserve and initialize a slot to hold a SlabPage.
Note that all SlabPages have the same size, independently of the const
generic parameter N. The new function propagates the generic parameter
to reduce the amount of casts needed.

Add as well a deallocate_page_slot() for symmetry, which both ensures
the type of the passed pointer is correct, and that the data page
belonging to the SlabPage is freed.

Signed-off-by: Carlos López <clopez@suse.de>