mm/alloc: remove SlabPage::free
Remove the free field, and instead track the number of free chunks in
the slab by counting zeroed bits in the bitmap, which can typically be
achieved with specialized instructions.

This change reduces the size of SlabPage from 40 to 32 bytes, increasing
the amount of SlabPage objects that can be stored in a single page from
102 to 128 (or around 25% more), meaning that less slab pages need to be
allocated to store SlabPage objects. This adds the extra cost of an extra
memory access to count the number of free slots in the slab page.

Signed-off-by: Carlos López <clopez@suse.de>