cpu/irq_state: make IrqState Sync
Make the type Sync to signify that it is suitable for use in the PerCpu.
Technically, the type cannot produce undefined behavior even under
concurrent usage, so this is correct. The type was marked !Sync to warn
users not to use it in a concurrent manner, as the type will not operate
correctly. Add a docstring to make this point it clear.

Signed-off-by: Carlos López <clopez@suse.de>