build: keep ELF symbol information in debug builds
When performing debug builds, keep symbol information sections in the
ELF file (`.symtab` and `.strtab`) by passing `--strip-debug` to obcopy.
This still removes pure debug sections (`.debug*`), but keeps `.symtab`
and `.strtab`, which will later be used to construct symbolized
stacktraces.

This increases the IGVM file size by around 20% to 25%, which is
why it is only enabled by default for debug builds. In the future, a
configuration option can be added to control this independently of the
type of build.

Signed-off-by: Carlos López <clopez@suse.de>