stage2: parse ELF symbol information
Now that the ELF parser keeps track of the `.symtab` and `.strtab`
sections, and that we have defined how to pass this information into the
main SVSM kernel, retrieve them and prepare heap-allocated regions to
feed into the SVSM.

To do this, introduce PageCursor, a struct to flexibly allocate pages
on demand from the KernelHeap and construct the right memory regions.

If the ELF sections are not presents, the code will pass empty regions,
which the SVSM will handle gracefully without panicking.

Signed-off-by: Carlos López <clopez@suse.de>